add note.ts
