+++ src/lib/util.ts
@@ -1,19 +1,3 @@
-function newNote(currentNotes: NoteState[]): NoteState {
-  const maxId = currentNotes.reduce((a, n) => (a < n.id ? n.id : a), 0);
-  return { id: maxId + 1 };
-}
-
-function appendNewNote(setNotes: SetNotesFunc): void {
-  setNotes(prev => [...prev, newNote(prev)]);
-}
-
-function insertNewNoteBefore(setNotes: SetNotesFunc, id: number): void {
-  setNotes(prev => {
-    const i = prev.findIndex(n => n.id === id);
-    return [...prev.slice(0, i), newNote(prev), ...prev.slice(i, prev.length)];
-  });
-}
-
 // 現在時刻を YYYYMMDD_HH-mm-ss の形式で返す関数
 function now(): string {
   const now = new Date();
@@ -41,5 +25,4 @@ const download: () => void = () => {
   a.dispatchEvent(new MouseEvent("click"));
 };
 
-export { makeGraph } from "./graph";
-export { appendNewNote, insertNewNoteBefore, download };
+export { download };
